Example of three of four reasons from official Drools Expert guide (http://www.jboss.org/drools/documentation) when you ought to use rule engine (Initial commit):
+ The problem is just too fiddly for traditional code.

The problem may not be complex, but you can't see a non-fragile way of
building a solution for it.

+ The problem is beyond any obvious algorithmic solution.

It is a complex problem to solve, there are no obvious traditional
solutions, or basically the problem isn't fully understood.

+ The logic changes often

The logic itself may even be simple but the rules change quite often. In
many organizations software releases are few and far between and
pluggable rules can help provide the "agility" that is needed and
expected in a reasonably safe way.